Removes DatabaseIdHasher and id_hashes table.
