--- apps/bfd-model/bfd-model-rif/src/main/resources/db/migration/V109__add_id_hashes.sql
@@ -1,6 +1,0 @@
--- Lookup table for DatabaseIdHasher persistent cache.
-CREATE TABLE id_hashes (
-    id varchar(80) NOT NULL,
-    hash varchar(64) NOT NULL,
-    CONSTRAINT id_hashes_key PRIMARY KEY (id)
-);
